move influx functions into stats.go and add cli app on main.go
